GSPI driver
set up a new function call of CMSIS api without slave device.
Now we dont have the clock or cs or anything on the analyser (as it should), but we still arrive in the receive function via interrupt (button)
We need to check if ambiq tied to silabs spitrs out anything.

Also we resoldered the pins and ow we do not have any problems on logic analyser